+++ src/pages/Cms/ContentEdit/helpers/CastPageFields/helpers/CastValueInput/CastValueInput.tsx
@@ -1,20 +1,19 @@
 import type { FC } from 'react';
 import { Input, Select } from '@forgedevstack/bear';
-import { InkEditor } from '@forgedevstack/ink';
 import { isStringValue } from '@utils';
 import { CAST_FIELD_TYPE } from '@pages/Cms/CastPages/CastPages.const';
 import {
   CAST_PAGE_VALUE_PREFIX,
-  CAST_RICH_MIN_HEIGHT_PX,
   CAST_TEXTAREA_ROWS,
   CAST_VALUE_INPUT_TYPE,
 } from '../../CastPageFields.const';
-import { parseCastSelectOptions } from '../../CastPageFields.utils';
+import { parseCastSelectOptions, isCastLongText } from '../../CastPageFields.utils';
 import type { CastValueInputProps } from './CastValueInput.types';
 
 export const CastValueInput: FC<CastValueInputProps> = (props) => {
   const { field, value, label, onValueChange } = props;
   const fieldId = `${CAST_PAGE_VALUE_PREFIX}${field.id}`;
+  const isLongText = isCastLongText(field.type);
 
   if (field.type === CAST_FIELD_TYPE.SELECT) {
     return (
@@ -33,19 +32,7 @@ export const CastValueInput: FC<CastValueInputProps> = (props) => {
     );
   }
 
-  if (field.type === CAST_FIELD_TYPE.RICH) {
-    return (
-      <InkEditor
-        value={value}
-        onChange={(next) => onValueChange(field.name, next)}
-        colorMode="light"
-        variant="document"
-        minHeight={CAST_RICH_MIN_HEIGHT_PX}
-      />
-    );
-  }
-
-  if (field.type === CAST_FIELD_TYPE.TEXTAREA) {
+  if (isLongText) {
     return (
       <Input
         id={fieldId}