+++ src/pages/Cms/ContentEdit/helpers/ContentTranslationWidget/index.ts
@@ -1,0 +1,5 @@
+export { ContentTranslationWidget } from './ContentTranslationWidget';
+export type {
+  ContentTranslationWidgetProps,
+  TranslationApplyParams,
+} from './ContentTranslationWidget.types';