+++ src/pages/Cms/TranslationsPages/helpers/TranslationsManager/helpers/TranslationAddBar/index.ts
@@ -1,0 +1,2 @@
+export { TranslationAddBar } from './TranslationAddBar';
+export type { TranslationAddBarProps } from './TranslationAddBar.types';