+++ src/pages/Cms/TranslationsPages/helpers/TranslationsManager/helpers/TranslationAddPage/TranslationAddPage.types.ts
@@ -1,0 +1,7 @@
+export type TranslationAddPageProps = {
+  titleValue: string;
+  titlePlaceholder: string;
+  addLabel: string;
+  onTitle: (value: string) => void;
+  onAdd: () => void;
+};