+++ src/pages/Cms/TranslationsPages/helpers/TranslationsManager/helpers/TranslationAddPage/index.ts
@@ -1,0 +1,2 @@
+export { TranslationAddPage } from './TranslationAddPage';
+export type { TranslationAddPageProps } from './TranslationAddPage.types';