+++ src/pages/Cms/TranslationsPages/helpers/TranslationsManager/index.ts
@@ -1,8 +1,17 @@
 export { TranslationsManager } from './TranslationsManager';
 export type { TranslationsManagerProps } from './TranslationsManager.types';
 export {
+  emptyBag,
+  pageKeyCount,
+  globalKeyCount,
+  seedIfEmpty,
+  addKeyToBag,
+  addPageToBag,
+  isTranslationKey,
+} from './TranslationsManager.utils';
+export {
   TRANSLATION_LOCALES,
   TRANSLATION_SEED,
   TRANSLATION_SCOPE,
+  TRANSLATION_GLOBAL_ID,
 } from './TranslationsManager.const';
-export { emptyBag, pageKeyCount, seedIfEmpty } from './TranslationsManager.utils';